requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.98.0
langchain==0.3.27
langchain-openai==0.3.28
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.1.1